cleanup clippy warns (#82)
* Refactor string formatting for improved readability and consistency

* Improve code clarity by renaming thread_id to _thread_id in LocalSpanManager

* Translate comments in entity.rs from Chinese to English for improved clarity and consistency

* Improve logging message formatting in setup and cleanup functions for consistency

* Refactor logging statements to use inline formatting for improved readability

* Refactor logging statements to use inline formatting for improved readability

* Refactor span statistics type for clarity and simplify authorization token extraction

+++ probing/cli/src/table.rs
@@ -68,7 +68,7 @@ pub fn render_dataframe(df: &DataFrame) {
     let mut table = Table::new(ncol, nrow);
 
     for (col, name) in df.names.iter().enumerate() {
-        table.put((0 as usize, col).into(), name.clone());
+        table.put((0_usize, col).into(), name.clone());
     }
 
     for (col, col_data) in df.cols.iter().enumerate() {
